# openspec/specs/openapi-contract.yaml
openapi: 3.0.3
info:
  title: Pharmacy API
  version: 1.0.0
paths:
  /api/productos:
    get:
      summary: Lista productos
      responses:
        '200': { description: Lista obtenida }
    post:
      summary: Crea un producto
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Producto' }
      responses:
        '201': { description: Producto creado }
        '400': { description: Datos inválidos }
  /api/productos/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer, format: int64 }
    get:
      summary: Obtiene un producto
      responses:
        '200': { description: Producto encontrado }
        '404': { description: Producto no encontrado }
    put:
      summary: Actualiza un producto
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Producto' }
      responses:
        '200': { description: Producto actualizado }
    delete:
      summary: Elimina un producto
      responses:
        '204': { description: Producto eliminado }
  /api/usuarios:
    get:
      summary: Lista usuarios
      responses:
        '200': { description: Lista obtenida }
    post:
      summary: Crea un usuario
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Usuario' }
      responses:
        '201': { description: Usuario creado }
        '400': { description: Datos inválidos }
        '409': { description: Nombre de usuario o email duplicado }
  /api/usuarios/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer, format: int64 }
    get:
      summary: Obtiene un usuario
      responses:
        '200': { description: Usuario encontrado }
        '404': { description: Usuario no encontrado }
    put:
      summary: Actualiza un usuario
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Usuario' }
      responses:
        '200': { description: Usuario actualizado }
        '404': { description: Usuario no encontrado }
    delete:
      summary: Elimina un usuario
      responses:
        '204': { description: Usuario eliminado }
        '404': { description: Usuario no encontrado }
components:
  schemas:
    Producto:
      type: object
      required: [nombre, precio, stock]
      properties:
        id: { type: integer, format: int64, readOnly: true }
        nombre: { type: string, maxLength: 120 }
        descripcion: { type: string, maxLength: 500 }
        precio: { type: number, format: double, minimum: 0 }
        stock: { type: integer, minimum: 0 }
    Usuario:
      type: object
      required: [nombreUsuario, email, contrasena, rol]
      properties:
        id: { type: integer, format: int64, readOnly: true }
        nombreUsuario: { type: string, maxLength: 120 }
        email: { type: string, format: email, maxLength: 120 }
        contrasena: { type: string, minLength: 6, writeOnly: true }
        rol: { type: string, enum: [ADMIN, ENCARGADO] }